+++ classes/local/participant/data_access/participant_vault_interface.php
@@ -127,11 +127,12 @@ interface participant_vault_interface {
     public function is_student_lessons_complete(int $userid, int $courseid, int $nextexercise);
 
     /**
-     * Updates a user's description field with a comment
+     * Updates a user's profile field with a value
      *
      * @param   int    $userid  The student user id
-     * @param   string $comment The next exercise id
-     * @return  bool    Whether the comment was updated or not.
+     * @param   string $field   The field to be updated
+     * @param   mixed  $value   The value to update to
+     * @return  bool            Whether the comment was updated or not.
      */
-    public function update_participant_comment(int $userid, string $comment);
+    public function update_participant_field(int $userid, string $field, $value);
 }